refactor(navigation): cleanup and refactoring
