+++ packages/frontend/src/components/accounts/account_selector/AccountSelector.js
@@ -53,13 +53,15 @@ const StyledContainer = styled.div`
     }
 `;
 
+const noop = () => {};
+
 export default ({
     signedInAccountId,
     availableAccounts,
     accountsBalances,
     getAccountBalance,
     onSelectAccount,
-    onSignInToDifferentAccount,
+    onSignInToDifferentAccount = noop,
     showBalanceInUSD
 }) => {
